feat: non-finite float inputs and a cortex-m0 f32 leg
Adds a descriptor-level `input_mode: nonfinite_sweep` that overwrites the first
four flat elements of a float input with NaN, -NaN, +Inf and -Inf, and gives
cortex-m0 the fp32_execution capability so the soft-float target runs the f32
float suite. Closes the two halves of #74.

Chose a generation mode over literal input vectors. The descriptors already
describe inputs declaratively (input_shape plus input_min/input_max) and never
carry sample data, so literal vectors would have meant hand-authoring an array
per shape and per dtype for each of the 16/17/64/128-element cases, and would
have bypassed the seeded sampler that keeps cases reproducible. A mode composes
with the existing shape and range fields instead, and the finite neighbours stay
as drawn, which is what distinguishes a kernel that corrupts the lanes around a
NaN from one that only mishandles the token.

Expected values come from the reference path each op already uses, so both
contract shapes fall out without special-casing: propagation (NaN in, NaN out)
and clamping, where the domain reduction bounds the result and ordinary
tolerance applies -- tanh(+-Inf) = +-1, sigmoid(+Inf) = 1, relu6(+Inf) = 6.

Serialization needed fixing first. format_float_literal had no non-finite path:
NaN rendered as the uninhabitable `nan.0f` and Inf raised a ValueError trying to
split an exponent off "inf". It now emits the C99 NAN/INFINITY macros, which are
compiler builtins and survive -Ofast: -ffinite-math-only lets the optimizer
assume operands are finite but does not stop a builtin materializing the bit
pattern. Verified on generated output by compiling to assembly for cortex-m0 at
-Ofast and reading the rodata -- 0x7fc00000, 0xffc00000, 0x7f800000, 0xff800000
all reach the image intact. The generated headers now include math.h directly
because they are included ahead of anything else that would define the macros.

Two reference paths needed guarding. The f16 MVE tanh model cast a NaN LUT index
through uint16, which is undefined in numpy exactly as it is in C -- the numpy
analogue of the out-of-bounds read in ns-cmsis-nn#314. PReLU's float path drew
its input directly from self.rng rather than the shared sampler, so it would
have accepted input_mode and silently ignored it.

Adds 20 cases: tanh, sigmoid, relu, relu6, hardswish and leaky_relu across f32
and f16, hard_swish and prelu float, and Add/Mul float as propagation sanity.
The Add/Mul +-Inf lanes land on the activation clamp rather than on Inf
passthrough in f32, which is noted where the cases are defined.

cortex-m0 needed no build change -- CMakeLists already routes it to
-mfloat-abi=soft with no FPU -- but it now joins cortex-m4 in the f32 build and
run groups, which is why the routing tests move from asserting a per-cpu command
to a shared one. The fp32 rejection guard in config.py has no naturally
capability-less cpu left to exercise it, so its test now uses a stripped profile
rather than being dropped.

+++ helia_core_tester/fvp/cmake.py
@@ -31,7 +31,7 @@ def active_test_list(generated_tests_dir: Optional[Path]) -> Optional[Set[str]]:
     -- but ``None`` is reserved strictly for that "no manifest file" case.
 
     A manifest that exists but admitted zero cases (every descriptor
-    capability-skipped for this cpu, e.g. the float suite on cortex-m0) is a
+    capability-skipped for this cpu, e.g. the f16 float suite on cortex-m0) is a
     real, empty active list, not "no constraint" -- it must return ``set()``,
     not ``None``, or build/run would fall back to unfiltered discovery and
     happily prune nothing / run every stale ``.elf`` in the tree.
